requirements.txt
mcp>=1.0.0
google-auth>=2.0.0
google-auth-oauthlib>=1.0.0
google-api-python-client>=2.0.0
starlette>=0.27.0
uvicorn>=0.20.0
python-dotenv>=1.0.0
